+++ backend/canisters/community/api/src/queries/mod.rs
@@ -1,6 +1,5 @@
-pub mod api_key;
-pub mod c2c_bot_api_key;
 pub mod c2c_bot_channel_details;
+pub mod c2c_bot_community_summary;
 pub mod c2c_can_issue_access_token;
 pub mod c2c_events;
 pub mod c2c_events_by_index;