+++ backend/canisters/community/c2c_client/src/lib.rs
@@ -2,8 +2,8 @@ use canister_client::generate_c2c_call;
 use community_canister::*;
 
 // Queries
-generate_c2c_call!(c2c_bot_api_key);
 generate_c2c_call!(c2c_bot_channel_details);
+generate_c2c_call!(c2c_bot_community_summary);
 generate_c2c_call!(c2c_can_issue_access_token);
 generate_c2c_call!(c2c_events);
 generate_c2c_call!(c2c_events_by_index);