+++ backend/canisters/community/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Additional notification data for Firebase Cloud Messaging ([8113](https://github.com/open-chat-labs/open-chat/pull/8113))
+- Add `c2c_bot_community_summary` endpoint ([#8135](https://github.com/open-chat-labs/open-chat/pull/8135))
 
 ### Changed
 
@@ -17,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove `correlation_id` ([#8097](https://github.com/open-chat-labs/open-chat/pull/8097))
 - Push chat event notifications to bots ([#8109](https://github.com/open-chat-labs/open-chat/pull/8109))
 - Remove bot API Keys ([#8131](https://github.com/open-chat-labs/open-chat/pull/8131))
+
+### Fixed
+
+- Add missing decoding for community read permissions ([#8141](https://github.com/open-chat-labs/open-chat/pull/8141))
 
 ## [[2.0.1763](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1763-community)] - 2025-05-28